chore(ui): extend linecheck 700-line gate to the ui crate
Splits the four ui/src files that exceeded 700 lines (routines.rs,
routines_tests.rs, main.rs, overview.rs) into smaller sibling modules,
mirroring the src/routines/ directory-module convention, then adds
ui/src to the pre-push line-count gate so it's enforced there too.
