+++ .changeset/ui-linecheck-700.md
@@ -1,0 +1,5 @@
+---
+"moadim": patch
+---
+
+chore(ui): extend the 700-line linecheck gate to the `ui` crate
